Review fixes
